requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
gradio>=4.0.0
python-doctr[torch]>=0.9.0
scikit-learn>=1.3.0
scikit-image>=0.20.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
jupyter>=1.0.0
notebook>=7.0.0
